% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvert.R
\name{plot.TempConvert}
\alias{plot.TempConvert}
\title{plot.TempConvert}
\usage{
\method{plot}{TempConvert}(x, metoda, ...)
}
\arguments{
\item{x}{temperatura}

\item{metoda}{wartosc  wartosc typu: "ctok", "ftoc","ctof","ktoc","ftok","ktof"}

\item{...}{inne argumenty}
}
\description{
Funkcja zwraca wykres w postaci Bar Plotu dla temperatur
}
\author{
Bartlomiej Kalata
}
