% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{print.TempConvert}
\alias{print.TempConvert}
\title{Przyrosty Print}
\usage{
\method{print}{TempConvert}(x, ...)
}
\arguments{
\item{x}{temperatura}

\item{...}{inne argumenty}
}
\description{
Funkcja wypisuje wprowadzone i przekonwertowane temperatury
}
\author{
Bartlomiej Kalata
}
