% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvert.R
\name{summary.TempConvert}
\alias{summary.TempConvert}
\title{Konwersja Temperatury Summary}
\usage{
\method{summary}{TempConvert}(x, metoda, ...)
}
\arguments{
\item{x}{temperatura}

\item{metoda}{wartosc  wartosc typu: "ctok", "ftoc","ctof","ktoc","ftok","ktof"}

\item{...}{inne argumenty}
}
\description{
Funkcja zwraca podstawowe informacje o obiekcie zwracanym z funkcji
}
\author{
Bartlomiej Kalata Piotr Pamula
}
