\name{TempConvert-package}
\alias{TempConvert-package}
\alias{TempConvert}
\docType{package}
\title{
\packageTitle{TempConvert}
}
\description{
\packageDescription{TempConvert}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TempConvert}
\packageIndices{TempConvert}
Deskrypcja
}
\author{
\packageAuthor{TempConvert}

Maintainer: \packageMaintainer{TempConvert}
}
\references{
ref
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
###
}
