% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{TempConvert}
\alias{TempConvert}
\title{Projekt konwertujacy temperatury}
\usage{
TempConvert(x, metoda, ...)
}
\arguments{
\item{x}{numeryczna wartosc}

\item{metoda}{wartosc  wartosc typu: "ctok", "ftoc","ctof","ktoc","ftok","ktof"}

\item{...}{inne argumenty}
}
\value{
Funkcja zwaraca w wyniku przekonwertowana temperature jako wartosc numeryczna
}
\description{
Funkcja zwraca przekonwertowane temperatury wykorzystujac klasyczne przeliczniki. Skala Celsjusza, Farenheita, Kelwina.
}
\author{
Bartlomiej Kalata
}
