% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot.TempConvert}
\alias{plot.TempConvert}
\title{plot.TempConvert}
\usage{
\method{plot}{TempConvert}(x, ...)
}
\arguments{
\item{x}{temperatura}

\item{...}{inne argumenty}
}
\description{
Funkcja zwraca wykres w postaci Bar Plotu dla temperatur
}
\author{
Bartlomiej Kalata
}
