% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvert.R
\name{print.TempConvert}
\alias{print.TempConvert}
\title{Przyrosty Print}
\usage{
\method{print}{TempConvert}(x, metoda, ...)
}
\arguments{
\item{x}{temperatura}

\item{metoda}{wartosc  wartosc typu: "ctok", "ftoc","ctof","ktoc","ftok","ktof"}

\item{...}{inne argumenty}
}
\description{
Funkcja wypisuje wprowadzone i przekonwertowane temperatury
}
\author{
Bartlomiej Kalata Piotr Pamula
}
