% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempConvert.R
\name{TempConvert.default}
\alias{TempConvert.default}
\title{TempConvert}
\usage{
\method{TempConvert}{default}(x, metoda = "ctok", ...)
}
\arguments{
\item{x}{numeryczna wartosc}

\item{metoda}{wartosc  wartosc typu: "ctok", "ftoc","ctof","ktoc","ftok","ktof"}

\item{...}{inne argumenty}
}
\value{
Funkcja zwaraca w wyniku przekonwertowana temperature jako wartosc numeryczna
}
\description{
Funkcja zwraca przekonwertowane temperatury wykorzystujac klasyczne przeliczniki. Skala Celsjusza, Farenheita, Kelwina.
}
\author{
Bartlomiej Kalata Piotr Pamula
}
